+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [v20.4.0](https://github.com/lexicalunit/spellbot/releases/tag/v20.4.0) - 2026-05-26
+
 ### Added
 
 - Added a `/queues.json` endpoint.